backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
twilio
apscheduler
pydantic
bcrypt
python-jose[cryptography]
python-multipart
